+++ packages/sdk/snet/sdk/service_client.py
@@ -8,7 +8,7 @@ from rfc3986 import urlparse
 from eth_account.messages import defunct_hash_message
 
 import snet.sdk.generic_client_interceptor as generic_client_interceptor
-from snet.mpe.payment_channel_provider import PaymentChannelProvider
+from snet.sdk.mpe.payment_channel_provider import PaymentChannelProvider
 from snet.snet_cli.utils import RESOURCES_PATH, add_to_path
